feat(docker): Add container, upgrade dependencies.
This introduces the use of [Docker Compose][1] to help make
cross-platform development for engineers a _lot_ easier. It also allows
for normalization of the development environemnt to our production
instances.

This included the adding of a few scripts that capture a lot of the work
that _could_ be stored in the Dockerfile but was extracted for size.

This also upgraded a few dependencies and enabled for the use of
`execjs`.

[1]: https://docs.docker.com/compose/